Backend & Frontend Storage Refactor

+++ social-campus-server/Infrastructure/Repositories/CommentLikeRepository.cs
@@ -5,31 +5,27 @@ using Domain.Models.UserModel;
 using Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure.Repositories
+namespace Infrastructure.Repositories;
+
+public class CommentLikeRepository(
+    ApplicationDbContext context) : ICommentLikeRepository
 {
-    public class CommentLikeRepository(
-        ApplicationDbContext context) : ICommentLikeRepository
+    public async Task AddAsync(UserId userId, CommentId commentId)
     {
-        public async Task AddAsync(UserId userId, CommentId commentId)
-        {
-            CommentLike newCommentLike = new(userId, commentId);
+        CommentLike newCommentLike = new(userId, commentId);
 
-            await context.CommentLikes.AddAsync(newCommentLike);
-        }
+        await context.CommentLikes.AddAsync(newCommentLike);
+    }
 
-        public async Task DeleteAsync(UserId userId, CommentId commentId)
-        {
-            CommentLike? commentLike = await context.CommentLikes
-                .FirstOrDefaultAsync(pl => pl.CommentId == commentId && pl.UserId == userId);
-            if (commentLike != null)
-            {
-                context.CommentLikes.Remove(commentLike);
-            }
-        }
+    public async Task DeleteAsync(UserId userId, CommentId commentId)
+    {
+        var commentLike = await context.CommentLikes
+            .FirstOrDefaultAsync(pl => pl.CommentId == commentId && pl.UserId == userId);
+        if (commentLike != null) context.CommentLikes.Remove(commentLike);
+    }
 
-        public async Task<bool> IsLike(UserId userId, CommentId commentId)
-        {
-            return await context.CommentLikes.AnyAsync(pl => pl.CommentId == commentId && pl.UserId == userId);
-        }
+    public async Task<bool> IsLike(UserId userId, CommentId commentId)
+    {
+        return await context.CommentLikes.AnyAsync(pl => pl.CommentId == commentId && pl.UserId == userId);
     }
 }
